+++ src/components/com_search/domains/queries/node.php
@@ -42,7 +42,8 @@ class ComSearchDomainQueryNode extends AnDomainQueryDefault
             'node.leader_count',
             'node.parent_id',
             'node.parent_type',
-            'node.filename'
+            'node.filename',
+            'node.cover_filename'
         ));
     }