+++ src/media/lib_anahita/js/site.js
@@ -52,7 +52,6 @@
 //@depends anahita/plyr.js
 //@depends anahita/cover.js
 
-
 //@depends anahita/actions/vote.js
 //@depends anahita/actions/entity.js
 //@depends anahita/actions/postlink.js